removed some wanrings for gcc, which complains about no newline at end of file
fix for Stream Handling, if a channel is choosen by the player which already
contains a stream, this stream is deleted then.

very important fix for unix/linux pthreads. it seems as if it is not allowed to
wait for the end of a thread in a differnt thread than the one , who started this
thread. code is modified now in a way that this does not occur anymore.
now i am rid of all the <defunct> process entries for Musik, which were generated
for every song.
hopefully this fixes the phenomena too, which leads to 99% cpu usage and not
palying songs anymore.